\docType{data}
\name{daphnia}
\alias{daphnia}
\title{Sample Daphnia Data}
\format{data frame}
\source{
   http://cbr.washington.edu/analysis/vitality
  
  Anderson, J.J. (2000). "A vitality-based model relating stressors and environmental properties to organism survival." Ecological Monographs 70(3):445-470 (Figure 5)
  
  
}
\description{
Sample survival data for daphnia. Columns include "days" and "lx" (cumulative survival proportion by day).
}
\keyword{datasets}

