\name{ft.4p}
\alias{ft.4p}
\title{Density function for 3-parameter (r, s, u)}
\usage{
  ft.4p(xx, r, s, u)
}
\arguments{
  \item{xx}{age}

  \item{r}{r value}

  \item{s}{s value}

  \item{u}{u value}
}
\value{
  density
}
\description{
  This function is used in the calculation of the fitted intrinsic (\code{\link{mu.vd1.4p}}) and total (\code{\link{mu.vd.4p}}) mortality rate in the 4-parameter model.
}
\seealso{\code{\link{vft.4p}}, \code{\link{ft.6p}}}
