\name{ft.6p}
\alias{ft.6p}
\title{Density function for 2-parameters (r, s)}
\usage{
  ft.6p(xx, r, s)
}
\arguments{
  \item{xx}{age}

  \item{r}{r value}

  \item{s}{s value}
}
\value{
  density
}
\description{
  This function is used in the calculation of the fitted intrinsic (\code{\link{mu.vd1.6p}}) and total (\code{\link{mu.vd.6p}}) mortality rate in the 6-parameter model.
}
\seealso{\code{\link{vft.6p}}}

