\name{logLikelihood.2ps}
\alias{logLikelihood.2ps}
\title{Gives log likelihood of 2-process 4-parameter model}
\usage{
  logLikelihood.2ps(par, xx1, xx2, NNi)
}
\arguments{
  \item{par}{vector of parameter(r, s, lambda, beta)}

  \item{xx1}{xx1 vector}

  \item{xx2}{xx2 vector}

  \item{NNi}{survival fractions}
}
\value{
  log likelihood
}
\description{
  None
}

