\name{mu.vd}
\alias{mu.vd}
\title{Total mortality rate}
\usage{
  mu.vd(t, r, s, lambda, beta, gamma, alpha)
}
\arguments{
  \item{t}{age}

  \item{r}{r value}

  \item{s}{s value}

  \item{lambda}{lambda value}

  \item{beta}{beta value}

  \item{gamma}{gamma value}

  \item{alpha}{alpha value}
}
\value{
  Total mortality rate (?)
}
\description{
  None
}

