\name{mu.vd1.4p}
\alias{mu.vd1.4p}
\title{Intrinsic mortality rate for the 2-process 4-parameter vitality model}
\usage{
  mu.vd1.4p(x, r, s)
}
\arguments{
  \item{x}{age}

  \item{r}{r value}

  \item{s}{s value}
}
\value{
  Intrinsic age-specific mortality rates
}
\description{
  Gives the intrinsic age-specific mortality rates for a given set of \code{r} and \code{s}, the intrinsic parameters.
}
\seealso{\code{\link{mu.vd.4p}}, \code{\link{mu.vd2.4p}}}

