\name{mu.vd2.4p}
\alias{mu.vd2.4p}
\title{Extrinsic mortality rate for the 2-process 4-parameter vitality model}
\usage{
  mu.vd2.4p(x, r, lambda, beta)
}
\arguments{
  \item{x}{age}

  \item{r}{r value}

  \item{lambda}{lambda value}

  \item{beta}{beta value}
}
\value{
  Extrinsic age-specific mortality rates
}
\description{
  Gives the extrinsic age-specific mortality rates for a given set of \code{r} and the extrinsic parameters.
}
\seealso{\code{\link{mu.vd.4p}}, \code{\link{mu.vd1.4p}}}

