\name{mu.vd2}
\alias{mu.vd2}
\title{Intrinsic mortality rate}
\usage{
  mu.vd2(x, r, lambda, beta, gamma, alpha)
}
\arguments{
  \item{x}{age}

  \item{r}{r value}

  \item{lambda}{lambda value}

  \item{beta}{beta value}

  \item{gamma}{gamma value}

  \item{alpha}{alpha value}
}
\value{
  Intrinsic mortality rate (?)
}
\description{
  None
}

