\name{mu.vd4.6p}
\alias{mu.vd4.6p}
\title{Childhood extrinsic mortality rate for the 2-process 6-parameter vitality model}
\usage{
  mu.vd4.6p(x, gamma, alpha)
}
\arguments{
  \item{x}{age}

  \item{gamma}{gamma value}

  \item{alpha}{alpha value}
}
\value{
  Vector of childhood extrinsic age-specific mortality rates at ages \code{x}
}
\description{
  Gives the childhood extrinsic age-specific mortality rates for a given set of the childhood extrinsic parameters.
}
\seealso{\code{\link{mu.vd.6p}}, \code{\link{mu.vd1.6p}},  \code{\link{mu.vd2.6p}},  \code{\link{mu.vd3.6p}}}
