\name{plotting.2ps}
\alias{plotting.2ps}
\title{Plotting function for 2-process vitality model. 4-param}
\usage{
  plotting.2ps(r.final, s.final, lambda.final, beta.final,
    mlv, time, sfract, x1, x2, Ni, pplot, Iplot, Mplot,
    tlab, rc.data)
}
\arguments{
  \item{r.final}{r estimate}

  \item{s.final}{s estimate}

  \item{lambda.final}{lambda estimate}

  \item{beta.final}{beta estimate}

  \item{mlv}{TODO mlv}

  \item{time}{time vector}

  \item{sfract}{survival fraction}

  \item{x1}{Time 1}

  \item{x2}{Time 2}

  \item{Ni}{Initial population}

  \item{pplot}{Boolean. Plot cumulative survival fraction?}

  \item{Iplot}{Boolean. Plot incremental survival?}

  \item{Mplot}{Boolean. Plot mortality rate?}

  \item{tlab}{Character, label for time axis}

  \item{rc.data}{Booolean, right-censored data?}
}
\description{
  None.
}

