\docType{data}
\name{rainbow_trout_for_k}
\alias{rainbow_trout_for_k}
\title{Sample Rainbow Trout Data}
\format{matrix}
\source{
  http://cbr.washington.edu/analysis/vitality
}
\description{
Sample survival data for rainbow trout. Columns include "days" and "survival" (cumulative survival proportion by day).
}
\keyword{datasets}

