\name{SurvFn.2ps}
\alias{SurvFn.2ps}
\title{The cumulative survival distribution function for 2-process 4-parameter}
\usage{
  SurvFn.2ps(xx, r, s, lambda, beta)
}
\arguments{
  \item{xx}{vector of ages}

  \item{r}{r value}

  \item{s}{s value}

  \item{lambda}{lambda value}

  \item{beta}{beta value}
}
\value{
  vector of FF?
}
\description{
  None.
}

