\name{SurvFn.h.6p}
\alias{SurvFn.h.6p}
\title{Intrinsic cumulative survival distribution for 6 parameter model}
\usage{
  SurvFn.h.6p(xx, r, s)
}
\arguments{
  \item{xx}{vector of ages}

  \item{r}{r value}

  \item{s}{s value}
}
\value{
  intrinsic cumulative survival distribution
}
\description{
  Gives the intrinsic cumulative survival distribution at \code{xx}.
}
\details{For use in \code{\link{vitality.6p}}.}
\seealso{\code{\link{vitality.6p}}}

