\name{SurvFn.h}
\alias{SurvFn.h}
\title{Intrinsic cumulative survival distribution}
\usage{
  SurvFn.h(xx, r, s, u)
}
\arguments{
  \item{xx}{vector of ages}

  \item{r}{r value}

  \item{s}{s value}

  \item{u}{u value}
}
\value{
  Cumulative survival distribution
}
\description{
  None
}

