\name{survProbInc.2ps}
\alias{survProbInc.2ps}
\title{Calculates incremental survival probability for 2-process 4-parameter r, s, lambda, beta}
\usage{
  survProbInc.2ps(r, s, lambda, beta, xx1, xx2)
}
\arguments{
  \item{r}{r value}

  \item{s}{s value}

  \item{lambda}{lambda value}

  \item{beta}{beta value}

  \item{xx1}{xx1 vector}

  \item{xx2}{xx2 vector}
}
\value{
  Incremental survival probabilities.
}
\description{
  None
}

