\name{vft.4p}
\alias{vft.4p}
\title{Vectorized density function}
\usage{
  vft.4p(xx, r, s, u)
}
\arguments{
  \item{xx}{vector of ages}

  \item{r}{r value}

  \item{s}{s value}

  \item{u}{u value}
}
\value{
  vector of densities
}
\description{
  This function is used in the calculation of the fitted intrinsic (\code{\link{mu.vd1.4p}}) and total (\code{\link{mu.vd.4p}}) mortality rate in the 4-parameter model.
}
\seealso{\code{\link{ft.4p}}}