\name{vft.6p}
\alias{vft.6p}
\title{Vectorized density function}
\usage{
  vft.6p(xx, r, s)
}
\arguments{
  \item{xx}{vector of ages}

  \item{r}{r value}

  \item{s}{s value}
}
\value{
  vector of densities
}
\description{
  This function is used in the calculation of the fitted intrinsic (\code{\link{mu.vd1.6p}}) and total (\code{\link{mu.vd.6p}}) mortality rate in the 6-parameter model.
}
\seealso{\code{\link{ft.6p}}}