\name{indexFinder}
\alias{indexFinder}
\title{Finds the first value of a vector that is less than a value.}
\usage{
  indexFinder(x, val)
}
\arguments{
  \item{x}{Vector to serach}

  \item{val}{Threshhold}
}
\value{
  Gives the index of the first value of x that is <= val.
  returns -1 if no value satisfies the condition
}
\description{
  None
}

