\name{mu.vd1.6p}
\alias{mu.vd1.6p}
\title{Intrinsic mortality rate for the 2-process 6-parameter vitality model}
\usage{
  mu.vd1.6p(x, r, s)
}
\arguments{
  \item{x}{age}

  \item{r}{r value}

  \item{s}{s value}
}
\value{
  Vector of intrinsic age-specific mortality rates at age \code{x}
}
\description{
  Gives the intrinsic age-specific mortality rates for a given set of \code{r} and \code{s}, the intrinsic parameters.
}
\seealso{\code{\link{mu.vd.6p}}, \code{\link{mu.vd2.6p}}}

