\docType{data}
\name{swedish_females}
\alias{swedish_females}
\title{Swedish Female Mortality Data}
\format{A \code{data.frame} object}
\source{
  Human Mortality Database
}
\description{
  Period mortality data for Swedish females experiencing
  mortality in the year 2000. Columns follow standard
  life-table naming conventions.
}
\keyword{datasets}

